% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
Use R package \code{microbenchmark} to compare the performance of C functions (\code{gibbsR} and \code{vaccR}) and Cpp functions (\code{gibbsC} and \code{vaccC}).
}
\examples{
\dontrun{
tm1 <- microbenchmark::microbenchmark(
  rnR = gibbsR(100,10,22,2,2),
  rnC = gibbsC(100,10,22,2,2)
)
print(summary(tm1)[,c(1,3,5,6)])
}
}
