% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{EM}
\alias{EM}
\title{EM algorithm estimates the joint density function of the normal mixed model}
\usage{
EM(x, C, iter, mu0)
}
\arguments{
\item{x}{data}

\item{C}{a count, guess there’s a mixture of C Gaussian distributions}

\item{iter}{count for iteration times}

\item{mu0}{initialize mu, a matrix size of C*ncol(x), you must first plot x, and estimate mu0 based on the picture, otherwise it will go wrong}
}
\value{
a list contains mu and Sigma
}
\description{
EM algorithm estimates the joint density function of the normal mixed model
}
\examples{
\dontrun{
x<-as.matrix(datasets::faithful)
plot(x)
EM(x,2,20,matrix(c(2,50,4,80),2,2,byrow=T))
}
}
