% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbsC(N, thin, n, aa, bb)
}
\arguments{
\item{N}{the number of samples}

\item{thin}{the number of between-sample random numbers}

\item{n}{the self-selected fixed parameter 'n' in ex9.8}

\item{aa}{the self-selected fixed parameter 'a' in ex9.8}

\item{bb}{the self-selected fixed parameter 'b' in ex9.8}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
rnC <- gibbsC(100,10,22,2,2)
par(mfrow=c(2,1));
plot(rnC[,1],type='l')
plot(rnC[,2],type='l')
}
}
