\name{StatComp21064-package}
\alias{StatComp21064-package}
\alias{StatComp21064}
\docType{package}
\title{
  This package provides several functions for benchmarking R and Rcpp.
}
\description{
   R and Rcpp versions of gibbs sampler are provided in this package (\code{gibbsR} versus \code{gibbsC})
}
\author{
Ruizhi Li

Maintainer: Ruizhi Li <736892430@qq.com>
}
\keyword{ package }